Moved code responsible for dealing with bitmaps to androidshared
