+++ androidshared/build.gradle
@@ -62,6 +62,7 @@ dependencies {
     implementation Dependencies.androidx_fragment_ktx
     implementation Dependencies.androidx_preference_ktx
     implementation Dependencies.timber
+    implementation Dependencies.androidx_exinterface
 
     testImplementation project(':testshared')
     testImplementation Dependencies.junit
@@ -69,6 +70,9 @@ dependencies {
     testImplementation Dependencies.androidx_test_espresso_core
     testImplementation Dependencies.robolectric
     testImplementation Dependencies.mockito_kotlin
+
+    androidTestImplementation Dependencies.androidx_test_ext_junit
+    androidTestImplementation Dependencies.junit
 
     debugImplementation project(':fragmentstest')
 }
