+++ androidshared/src/androidTest/java/org/odk/collect/androidshared/bitmap/ImageFileUtilsTest.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.odk.collect.android.instrumented.utilities
+package org.odk.collect.androidshared.bitmap
 
 import android.content.Context
 import android.graphics.Bitmap
@@ -26,7 +26,6 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.odk.collect.android.utilities.ImageFileUtils
 import timber.log.Timber
 import java.io.File
 import java.io.IOException
@@ -50,7 +49,10 @@ class ImageFileUtilsTest {
         attributes[ExifInterface.TAG_ORIENTATION] = ExifInterface.ORIENTATION_ROTATE_90.toString()
         saveTestBitmapToFile(sourceFile.absolutePath, attributes)
         ImageFileUtils.copyImageAndApplyExifRotation(sourceFile, destinationFile)
-        val image = ImageFileUtils.getBitmap(destinationFile.absolutePath, BitmapFactory.Options())!!
+        val image = ImageFileUtils.getBitmap(
+            destinationFile.absolutePath,
+            BitmapFactory.Options()
+        )!!
 
         assertEquals(2, image.width)
         assertEquals(1, image.height)
@@ -64,7 +66,10 @@ class ImageFileUtilsTest {
         attributes[ExifInterface.TAG_ORIENTATION] = ExifInterface.ORIENTATION_ROTATE_270.toString()
         saveTestBitmapToFile(sourceFile.absolutePath, attributes)
         ImageFileUtils.copyImageAndApplyExifRotation(sourceFile, destinationFile)
-        val image = ImageFileUtils.getBitmap(destinationFile.absolutePath, BitmapFactory.Options())!!
+        val image = ImageFileUtils.getBitmap(
+            destinationFile.absolutePath,
+            BitmapFactory.Options()
+        )!!
 
         assertEquals(2, image.width)
         assertEquals(1, image.height)
@@ -78,7 +83,10 @@ class ImageFileUtilsTest {
         attributes[ExifInterface.TAG_ORIENTATION] = ExifInterface.ORIENTATION_ROTATE_180.toString()
         saveTestBitmapToFile(sourceFile.absolutePath, attributes)
         ImageFileUtils.copyImageAndApplyExifRotation(sourceFile, destinationFile)
-        val image = ImageFileUtils.getBitmap(destinationFile.absolutePath, BitmapFactory.Options())!!
+        val image = ImageFileUtils.getBitmap(
+            destinationFile.absolutePath,
+            BitmapFactory.Options()
+        )!!
 
         assertEquals(1, image.width)
         assertEquals(2, image.height)
@@ -92,7 +100,10 @@ class ImageFileUtilsTest {
         attributes[ExifInterface.TAG_ORIENTATION] = ExifInterface.ORIENTATION_UNDEFINED.toString()
         saveTestBitmapToFile(sourceFile.absolutePath, attributes)
         ImageFileUtils.copyImageAndApplyExifRotation(sourceFile, destinationFile)
-        val image = ImageFileUtils.getBitmap(destinationFile.absolutePath, BitmapFactory.Options())!!
+        val image = ImageFileUtils.getBitmap(
+            destinationFile.absolutePath,
+            BitmapFactory.Options()
+        )!!
 
         assertEquals(1, image.width)
         assertEquals(2, image.height)
@@ -105,7 +116,10 @@ class ImageFileUtilsTest {
     fun copyAndRotateImageNoExif() {
         saveTestBitmapToFile(sourceFile.absolutePath, null)
         ImageFileUtils.copyImageAndApplyExifRotation(sourceFile, destinationFile)
-        val image = ImageFileUtils.getBitmap(destinationFile.absolutePath, BitmapFactory.Options())!!
+        val image = ImageFileUtils.getBitmap(
+            destinationFile.absolutePath,
+            BitmapFactory.Options()
+        )!!
 
         assertEquals(1, image.width)
         assertEquals(2, image.height)
